removal of uneccessery castings

+++ WindowsFeatures.cpp
@@ -43,7 +43,7 @@ uint32_t da::WindowsFeatures::GetThreadCount(uint64_t windowWidth, uint64_t wind
 
 bool da::WindowsFeatures::IsEnoughFreeMemory(uint64_t windowWidth, uint64_t windowHeight, uint64_t sizeOfItem)
 {
-	uint64_t TakenMemory = uint64_t(windowWidth) * uint64_t(windowHeight) * sizeOfItem / da::KB;
+	uint64_t TakenMemory = windowWidth * windowHeight * sizeOfItem / da::KB;
 	if (TakenMemory > da::WindowsFeatures::GetFreeMemoryInKB())
 	{
 		std::cout << " Not enough free memory for this mesh size" << std::endl;
